extracted parse_profiles() function into ProfileParser class; extracted map_record_to_json() function into GnomadVcfRecordParser class; extracted parse_site_quality_metrics() function into SiteQualityMetricsParser class and made an abstract class over it; moved the common parser classes into gnomad_common_parser module
